Accelerometer control
Yogu can be controlled via the accelerometer of the android device, if available.

Early-development state! It's quite difficult to play right now.

+++ src/de/yogularm/Player.java
@@ -44,7 +44,7 @@ public class Player extends Character {
 		return collectedCoins;
 	}
 	
-	public void setDirection(int direction) {
+	public void setDirection(float direction) {
 		setWalkSpeed(direction * Config.PLAYER_SPEED);
 	}
